% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{get_table(db_name, tbl_name)}
\usage{
get_table(db_name = "Elmer", schema, tbl_name)
}
\arguments{
\item{db_name}{String.  The name of the database to run the query against.  Should be "Elmer" or "ElmerGeo".  Default = "Elmer".}

\item{schema}{String.  The name of the schema that the table exists in.  No default.}

\item{tbl_name}{String.  The name of the table to be retrieved.  No default.}
}
\value{
A data frame.
}
\description{
Retrieve a dataset defined by a table name.
}
\examples{
get_table('chas', 'tenure_dim')

}
