% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_feature_classes.R
\name{list_feature_classes}
\alias{list_feature_classes}
\title{list_feature_classes(feature_dataset, feature_class)}
\usage{
list_feature_classes(feature_dataset = NULL, feature_class = NULL)
}
\arguments{
\item{feature_dataset}{String.  If supplied, this filters the list to just those feature classes in the feature dataset}

\item{feature_class}{String.  If supplied, this filters the list to just the named one.  This can be useful if you know the name of the feature class but not the geography type or the feature dataset that it resides in.}
}
\value{
A dataframe listing the layer names along with the corresponding feature datasets and geometry types..  Defaults to NULL.
}
\description{
list the feature classes available in ElmerGeo
}
\examples{
list_feature_classes()
list_feature_classes(feature_dataset = "equity")

}
