% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_layers.R
\name{list_layers}
\alias{list_layers}
\title{list_layers(feature_dataset)}
\usage{
list_layers(feature_dataset = NULL)
}
\arguments{
\item{feature_dataset}{String.  If supplied, this filters the list to just those layers in the feature dataset}
}
\value{
A dataframe listing the layer names along with the corresponding feature datasets and geometry types..  Defaults to NULL.
}
\description{
list the layers available in ElmerGeo
}
\examples{
list_layers()
list_layers(feature_dataset = "equity")

}
