% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_recordsets.R
\name{list_recordsets}
\alias{list_recordsets}
\title{list_recordsets(schema_name, recordset_type)}
\usage{
list_recordsets(schema_name = "", include_base_tables = FALSE)
}
\arguments{
\item{schema_name}{(string, optional) If supplied, this limits the returned list to only those tables and views in schema_name.}

\item{include_base_tables}{(TRUE|FALSE, optional) If false, return only the views (no tables).  False is the default.}
}
\value{
a list of tables, in the form of a dataframe
}
\description{
return a list of tables and views available in PSRC's in-house data warehouse
}
\examples{
list_recordsets()

list_recordsets(schema_name="HHSurvey")

list_recordsets(schema_name="HHSurvey", include_base_tables=TRUE)

}
