% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_query.R
\name{send_query}
\alias{send_query}
\title{send_query(db_name, sql)}
\usage{
send_query(db_name = "Elmer", sql)
}
\arguments{
\item{db_name}{String.  The name of the database to run the query against.  Should be "Elmer" or "ElmerGeo".  Default = "Elmer".}

\item{sql}{String.  The SQL command to send to <db_name>.}
}
\description{
Execute a SQL command to the database
}
\details{
Intended for update, merge or other commands that do not retrieve records. For retrieval, use get_query.
}
