% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_read_query.R
\name{st_read_query}
\alias{st_read_query}
\title{st_read_query(sql, ...)}
\usage{
st_read_query(sql, db_name = "Elmer", project_to_wgs84 = TRUE)
}
\arguments{
\item{sql}{String.  The SQL command to send to <db_name>.}

\item{db_name}{String.  The name of the database to run the query against.  Should be "Elmer" or "ElmerGeo".  Default = "Elmer".}

\item{project_to_wgs84}{(TRUE/FALSE) If TRUE then deliver the output in WGS84 projection, otherwise NAD84 / WA State Plane North.  Defaults to TRUE.}
}
\value{
object of class sf
}
\description{
Read geometric data to simple features with a query
Allows you to return part of a spatial table
For an entire spatial table from either Elmer or ElmerGeo, use st_read_elmergeo(layer_name, ...)
}
