% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_query.R
\name{stage_table}
\alias{stage_table}
\title{stage_table(df, table_name)}
\usage{
stage_table(df, table_name = name(df))
}
\arguments{
\item{df}{dataframe - the data object to write}

\item{table_name}{string - the name you want given the database table}
}
\description{
Write a dataframe as a table within the Elmer.stg schema
}
