% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query.R
\name{get_query}
\alias{get_query}
\title{get_query(db_name, sql)}
\usage{
get_query(db_name = "Elmer", sql)
}
\arguments{
\item{db_name}{String.  The name of the database to run the query against.  Should be "Elmer" or "ElmerGeo".  Default = "Elmer".}

\item{sql}{String.  The SQL command to send to <db_name>.}
}
\value{
A data frame.
}
\description{
Retrieve a dataset defined by a SQL string
}
\examples{
get_query("select * from chas.tenure_dim")
get_query(db_name = "ElmerGeo", sql = "select geoid10 from dbo.TRACT2010")

}
