% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_read_elmergeo.R
\name{st_read_elmergeo}
\alias{st_read_elmergeo}
\title{st_read_elmergeo(layer_name, ...)}
\usage{
st_read_elmergeo(layer_name, schema_name = "dbo")
}
\arguments{
\item{layer_name}{The name of the feature layer or geodatabase table}

\item{schema_name}{The name of the schema that layer_name exists in.  Defaults to "dbo", the standard schema for ElmerGeo tables}
}
\value{
object of class sf, projected in WSG84 coordinate system
}
\description{
read simple features from PSRC's in-house geodatabase #'
This function is preferable to sf::st_read() when pulling
from Elmer or ElmerGeo, because it avoids some incompatibilities
between SQL Server's data types and R's.
}
\note{
If the layer has been set up as an ESRI versioned layer in the geodatabase, this function returns the versioned view (which exists in SQL Server with a "_evw" suffix).  If it has not been set up that way, it returns the base table.
}
\examples{
st_read_elmergeo("COUNTY_BACKGROUND")

}
